saved all features as files, added support to quiz name , time... , changed the button to create quiz and edited the main page bar
